pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arbiter"
version = "0.1.0"
description = "Streaming LLM evaluation framework with semantic comparison and composable metrics"
authors = [{name = "Evan Volgas", email = "evan.volgas@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.14"
keywords = ["ai", "llm", "evaluation", "testing", "metrics", "semantic-comparison"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]

# Core dependencies - minimal for basic functionality
dependencies = [
    "pydantic>=2.11.0",
    "pydantic-ai>=0.1.0",
    "pymilvus>=2.4.0",  # Milvus vector database for semantic comparison
    "httpx>=0.28.0",
    "openai>=1.82.0",
    "python-dotenv>=1.0.1",
    "aiofiles>=24.1.0",
    "logfire>=3.21.0",
    "numpy>=1.24.0",  # For vector operations
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
# Model providers
anthropic = ["anthropic>=0.50.0"]
gemini = ["google-generativeai>=0.8.5"]

# Streaming support
streaming = ["bytewax>=0.20.0"]

# Storage backends
redis = ["redis>=5.0.0"]

# Development tools
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.6.0",  # Parallel test execution

    # Code quality
    "black>=24.2.0",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
    "types-redis>=4.6.0",
    "types-aiofiles>=24.1.0",

    # Interactive development
    "ipython>=8.0.0",
    "rich>=13.7.0",
    "ipdb>=0.13.13",

    # Documentation
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",

    # Development utilities
    "pre-commit>=3.7.0",
    "watchdog>=4.0.0",
    "psutil>=5.9.0",
    "setuptools>=70.0.0",
]

# All features
all = ["arbiter[anthropic,gemini,streaming,redis,dev]"]

[tool.setuptools.packages.find]
where = ["."]
include = ["arbiter*"]

[tool.setuptools.package-data]
arbiter = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
]

ignore = [
    "E501",   # line too long (handled by formatter)
]

fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["arbiter"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = ["--strict-markers", "--cov=arbiter", "--cov-report=term-missing"]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "performance: marks tests as performance tests"
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
pretty = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_reexport = true
strict_optional = true
strict_equality = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["pymilvus", "pymilvus.*", "logfire", "bytewax", "bytewax.*"]
ignore_missing_imports = true
